pyproject.toml
[project]
name = "logo-detection-service"
version = "0.1.0"
description = "Logo detection and classification service using GroundingDINO, Jina, Postgres, Qdrant, and S3-compatible storage."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "boto3>=1.35.0",
    "aiohttp>=3.13.4",
    "fastapi>=0.115.0",
    "Pillow>=10.4.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "qdrant-client>=1.12.0",
    "scalar-fastapi>=1.4.1",
    "SQLAlchemy>=2.0.36",
    "tenacity>=9.0.0",
    "torch>=2.4.0",
    "transformers>=4.56.0",
    "uvicorn[standard]>=0.32.0",
    "torchvision>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.8",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["B", "E", "F", "I"]
